nptl: Add public __rseq_feature_size symbol
Exposing this symbol allows applications wishing to use rseq features
which are part of the extensible rseq ABI like 'node_id' and 'mm_cid' to
test the two following conditions in a single load / conditional branch:

- rseq is registered
- the specific rseq feature is available

This is useful as rseq is expected to be used in hot paths.

This variable is either zero (if restartable sequence registration
failed or has been disabled) or the size of the available restartable
sequence features.

Change-Id: I2ecde62fc81d12416aa99bb404fe03b215ee25f6
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
---
Changes since v7:
- Fix sorting of symbols in abilist files
Changes since v8:
- Remove superfluous attributes on externs
- Clarify where the magic '20' initial feature size comes from