elf: Add rseq auxvals
Get the rseq feature size and alignment requirement from the auxiliary
vector for use inside the dynamic loader.

This will be used in the TLS block allocator to compute the size and
alignment of the rseq area block for the extended ABI support and at
thread creation to determine the size of the rseq registration.

Change-Id: I30ba5f865cdfef5c36f67fcc662fed3e3831ef74
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
---
Open questions:
- Should '_dl_rseq_feature_size' and '_dl_rseq_align' be in GLRO? If not
  where else?
- Since RSEQ is Linux specific, should the changes in 'elf/' and
  'sysdeps/generic/' be moved to 'sysdeps/unix/sysv/linux/'? If so, I'm
  unsure how it would work for GLRO variables.