nptl: Move the rseq area to the 'extra TLS' block
Move the rseq area to the newly added 'extra TLS' block, this is the
last step in adding support for the rseq extended ABI. The size of the
rseq area is now dynamic and depends on the rseq features reported by
the kernel through the elf auxiliary vector. This will allow
applications to use rseq features past the 32 bytes of the original rseq
ABI as they become available in future kernels.

Change-Id: Ia54cdb99a5d60f5fefe50e26fd1ccddecc479580
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>